The 'aalt' table for Latin contains only bengali characters. [renamed accordingly, and put in script 'beng' rather than 'latn']
The table "Ligature Substitution for Devanagari lookup 46" is really a 'reph'.   Re-named accordingly.

The table "'vatu' Vattu Variants lookup 24" is in script 'DFLT', which is wrong.  The characters are from Gujarati.  But there is already such a table for Gujarati.  Changed name and script accordingly.

Got rid of Single Substitution table.